update serilog, update rollingfile sink to non-deprecated sink
There are no functional changes here, just an update to remove a
deprecated package from the project, and to jump to the newest stable
release of serilog and the serilog file sink.

RollingFile was deprecated, and the File sink has a built-in rotation
option, which this switches to using.

Tested: built and run locally.  Worked as expected.  Output files were
still `{level}-{date}.log`, eg: `Error-20201005.log`