core: be more robust when launched before the PoE client starts
When ExileAPI was launched before the PoE client it ended up triggering
an exception every tick, flooding the error logs.  While these were
caught and the process continued, it doesn't cost much to avoid the null
dereference.

This fixes all the pre-launch, and during-launch exceptions for me,
making it easier to handle some testing scenarios.