Item Base Component updates
Move Base component data to use a struct for reading, not one-byte
remote reads.  Not a huge performance improvement, but not terrible,
given the frequency of access.

Found the correct offset for the influence flags, and verified it with
a range of influenced and non-influenced items.

Sadly, I failed to find the correct offset for the corrupted flag before
running out of patience.  Added a comment about what I found.