Update Offsets.cs
The start offset for the Area Change pattern was causing the scanner to start beyond where the offset is actually found. The File Root offset keeps getting pushed further from its start offset. Adjusted both to match the same scan time as the game state pattern (7ms each on my machine). This may be too aggressive and should be kept in mind with future patches.